refactor: prepare mulit-command cli
