+++ packages/cli/src/cli.js
@@ -1,60 +1,35 @@
 #!/usr/bin/env node
-import chalk from "chalk";
-import meow from "meow";
-import ora from "ora";
-import patternplate from ".";
+const chalk = require("chalk");
+const meow = require("meow");
 
 const cli = meow(
   `
-	Usage
-	$ patternplate [command=start] [options]
+  Usage
+  $ patternplate [command=start] [options]
 
-	Commands
-	  start   - start a patternplate instance in cwd
-	  help    - show this help
+  Commands
+    start   - Start a patternplate instance in cwd
+    help    - Show this help
 
-	Examples
-	  $ patternplate
-	`
+  Start options
+    --port  - Port to start patternplate server on, defaults to 1337, $PORT
+    --open  - Open the interface in system default browser
+
+  Examples
+    $ patternplate
+`
 );
 
 async function main({ input, flags }) {
   const [command] = input;
 
-  if (command === "help") {
-    cli.showHelp(0);
-    return;
-  }
-
-  const spinner = ora("Starting patternplate server").start();
-
-  const port = selectPort(cli.flags);
-
-  try {
-    const app = await patternplate({
-      port,
-      cwd: flags.cwd || process.cwd()
-    });
-
-    spinner.text = `Started on http://localhost:${app.port}`;
-    spinner.succeed();
-    app.subscribe(message => {
-      if (message.type === "error" && message.payload && typeof message.payload.message === "string") {
-        spinner.text = message.payload.message;
-        spinner.fail();
-      }
-    });
-  } catch (err) {
-    switch (err.code) {
-      case "EADDRINUSE":
-        spinner.text = `Starting patternplate server failed`;
-        spinner.fail();
-        err.message = `Server could not be started, free the port: ${err.message}`;
-        err.patternplate = true;
-        throw err;
-      default:
-        throw err;
-    }
+  switch (command) {
+    case 'help':
+      return cli.showHelp(0);
+    case 'start':
+    default:
+      const start = require("./commands/start");
+      return start({input, flags});
   }
 }
 
@@ -73,14 +48,3 @@ process.on("unhandledRejection", reason => {
   throw reason;
 });
 
-function selectPort(flags) {
-  if (!isNaN(Number(flags.port))) {
-    return Number(flags.port);
-  }
-
-  if (!isNaN(Number(process.env.PORT))) {
-    return (process.env.PORT);
-  }
-
-  return 1337;
-}
